#622:Moved course zip functions to a helper module
